requirements.txt
pyttsx3==2.90
pystray==0.19.5
keyboard==0.13.5
pyperclip==1.8.2
pillow==10.4.0
python-dotenv==1.0.0